extract testing policy snapshot microcrate
