+++ Cargo.toml
@@ -72,6 +72,7 @@ members = [
   "crates/bitnet-testing-profile",
   "crates/bitnet-testing-scenarios",
   "crates/bitnet-testing-policy-core",
+  "crates/bitnet-testing-policy-snapshot-core", # SRP: policy snapshot resolution helpers
   "crates/bitnet-testing-scenarios-core",
   "crates/bitnet-testing-scenarios-profile-core",
   "crates/bitnet-testing-policy", 
@@ -160,6 +161,7 @@ default-members = [
   "crates/bitnet-feature-matrix",
   "crates/bitnet-server-health-types-core",
   "crates/bitnet-testing-policy-core",
+  "crates/bitnet-testing-policy-snapshot-core", # SRP: policy snapshot resolution helpers
   "crates/bitnet-testing-scenarios-core",
   "crates/bitnet-runtime-profile-core",
   # SRP microcrate extractions (phase-6)
