+++ crates/bitnet-testing-policy-core/Cargo.toml
@@ -14,50 +14,51 @@ include = ["src/**", "Cargo.toml"]
 
 [dependencies]
 bitnet-testing-profile = { path = "../bitnet-testing-profile", version = "0.2.1-dev" }
+bitnet-testing-policy-snapshot-core = { path = "../bitnet-testing-policy-snapshot-core", version = "0.2.1-dev" }
 bitnet-testing-scenarios-core = { path = "../bitnet-testing-scenarios-core", version = "0.2.1-dev" }
 
 [features]
 default = []
-cpu = ["bitnet-testing-profile/cpu"]
-gpu = ["bitnet-testing-profile/gpu"]
-cuda = ["bitnet-testing-profile/cuda"]
-inference = ["bitnet-testing-profile/inference"]
-kernels = ["bitnet-testing-profile/kernels"]
-tokenizers = ["bitnet-testing-profile/tokenizers"]
-quantization = ["bitnet-testing-profile/quantization"]
-cli = ["bitnet-testing-profile/cli"]
-server = ["bitnet-testing-profile/server"]
-ffi = ["bitnet-testing-profile/ffi"]
-python = ["bitnet-testing-profile/python"]
-wasm = ["bitnet-testing-profile/wasm"]
-crossval = ["bitnet-testing-profile/crossval"]
-trace = ["bitnet-testing-profile/trace"]
-iq2s-ffi = ["bitnet-testing-profile/iq2s-ffi"]
-cpp-ffi = ["bitnet-testing-profile/cpp-ffi"]
-fixtures = ["bitnet-testing-profile/fixtures", "bitnet-testing-scenarios-core/fixtures"]
-reporting = ["bitnet-testing-profile/reporting"]
-trend = ["bitnet-testing-profile/trend"]
-integration-tests = ["bitnet-testing-profile/integration-tests"]
-runtime-profile-cpu = ["bitnet-testing-profile/runtime-profile-cpu"]
-runtime-profile-gpu = ["bitnet-testing-profile/runtime-profile-gpu"]
-runtime-profile-cuda = ["bitnet-testing-profile/runtime-profile-cuda"]
-runtime-profile-inference = ["bitnet-testing-profile/runtime-profile-inference"]
-runtime-profile-kernels = ["bitnet-testing-profile/runtime-profile-kernels"]
-runtime-profile-tokenizers = ["bitnet-testing-profile/runtime-profile-tokenizers"]
-runtime-profile-quantization = ["bitnet-testing-profile/runtime-profile-quantization"]
-runtime-profile-cli = ["bitnet-testing-profile/runtime-profile-cli"]
-runtime-profile-server = ["bitnet-testing-profile/runtime-profile-server"]
-runtime-profile-ffi = ["bitnet-testing-profile/runtime-profile-ffi"]
-runtime-profile-python = ["bitnet-testing-profile/runtime-profile-python"]
-runtime-profile-wasm = ["bitnet-testing-profile/runtime-profile-wasm"]
-runtime-profile-crossval = ["bitnet-testing-profile/runtime-profile-crossval"]
-runtime-profile-trace = ["bitnet-testing-profile/runtime-profile-trace"]
-runtime-profile-iq2s-ffi = ["bitnet-testing-profile/runtime-profile-iq2s-ffi"]
-runtime-profile-cpp-ffi = ["bitnet-testing-profile/runtime-profile-cpp-ffi"]
-runtime-profile-fixtures = ["bitnet-testing-profile/runtime-profile-fixtures"]
-runtime-profile-reporting = ["bitnet-testing-profile/runtime-profile-reporting"]
-runtime-profile-trend = ["bitnet-testing-profile/runtime-profile-trend"]
-runtime-profile-integration-tests = ["bitnet-testing-profile/runtime-profile-integration-tests"]
+cpu = ["bitnet-testing-profile/cpu", "bitnet-testing-policy-snapshot-core/cpu"]
+gpu = ["bitnet-testing-profile/gpu", "bitnet-testing-policy-snapshot-core/gpu"]
+cuda = ["bitnet-testing-profile/cuda", "bitnet-testing-policy-snapshot-core/cuda"]
+inference = ["bitnet-testing-profile/inference", "bitnet-testing-policy-snapshot-core/inference"]
+kernels = ["bitnet-testing-profile/kernels", "bitnet-testing-policy-snapshot-core/kernels"]
+tokenizers = ["bitnet-testing-profile/tokenizers", "bitnet-testing-policy-snapshot-core/tokenizers"]
+quantization = ["bitnet-testing-profile/quantization", "bitnet-testing-policy-snapshot-core/quantization"]
+cli = ["bitnet-testing-profile/cli", "bitnet-testing-policy-snapshot-core/cli"]
+server = ["bitnet-testing-profile/server", "bitnet-testing-policy-snapshot-core/server"]
+ffi = ["bitnet-testing-profile/ffi", "bitnet-testing-policy-snapshot-core/ffi"]
+python = ["bitnet-testing-profile/python", "bitnet-testing-policy-snapshot-core/python"]
+wasm = ["bitnet-testing-profile/wasm", "bitnet-testing-policy-snapshot-core/wasm"]
+crossval = ["bitnet-testing-profile/crossval", "bitnet-testing-policy-snapshot-core/crossval"]
+trace = ["bitnet-testing-profile/trace", "bitnet-testing-policy-snapshot-core/trace"]
+iq2s-ffi = ["bitnet-testing-profile/iq2s-ffi", "bitnet-testing-policy-snapshot-core/iq2s-ffi"]
+cpp-ffi = ["bitnet-testing-profile/cpp-ffi", "bitnet-testing-policy-snapshot-core/cpp-ffi"]
+fixtures = ["bitnet-testing-profile/fixtures", "bitnet-testing-scenarios-core/fixtures", "bitnet-testing-policy-snapshot-core/fixtures"]
+reporting = ["bitnet-testing-profile/reporting", "bitnet-testing-policy-snapshot-core/reporting"]
+trend = ["bitnet-testing-profile/trend", "bitnet-testing-policy-snapshot-core/trend"]
+integration-tests = ["bitnet-testing-profile/integration-tests", "bitnet-testing-policy-snapshot-core/integration-tests"]
+runtime-profile-cpu = ["bitnet-testing-profile/runtime-profile-cpu", "bitnet-testing-policy-snapshot-core/runtime-profile-cpu"]
+runtime-profile-gpu = ["bitnet-testing-profile/runtime-profile-gpu", "bitnet-testing-policy-snapshot-core/runtime-profile-gpu"]
+runtime-profile-cuda = ["bitnet-testing-profile/runtime-profile-cuda", "bitnet-testing-policy-snapshot-core/runtime-profile-cuda"]
+runtime-profile-inference = ["bitnet-testing-profile/runtime-profile-inference", "bitnet-testing-policy-snapshot-core/runtime-profile-inference"]
+runtime-profile-kernels = ["bitnet-testing-profile/runtime-profile-kernels", "bitnet-testing-policy-snapshot-core/runtime-profile-kernels"]
+runtime-profile-tokenizers = ["bitnet-testing-profile/runtime-profile-tokenizers", "bitnet-testing-policy-snapshot-core/runtime-profile-tokenizers"]
+runtime-profile-quantization = ["bitnet-testing-profile/runtime-profile-quantization", "bitnet-testing-policy-snapshot-core/runtime-profile-quantization"]
+runtime-profile-cli = ["bitnet-testing-profile/runtime-profile-cli", "bitnet-testing-policy-snapshot-core/runtime-profile-cli"]
+runtime-profile-server = ["bitnet-testing-profile/runtime-profile-server", "bitnet-testing-policy-snapshot-core/runtime-profile-server"]
+runtime-profile-ffi = ["bitnet-testing-profile/runtime-profile-ffi", "bitnet-testing-policy-snapshot-core/runtime-profile-ffi"]
+runtime-profile-python = ["bitnet-testing-profile/runtime-profile-python", "bitnet-testing-policy-snapshot-core/runtime-profile-python"]
+runtime-profile-wasm = ["bitnet-testing-profile/runtime-profile-wasm", "bitnet-testing-policy-snapshot-core/runtime-profile-wasm"]
+runtime-profile-crossval = ["bitnet-testing-profile/runtime-profile-crossval", "bitnet-testing-policy-snapshot-core/runtime-profile-crossval"]
+runtime-profile-trace = ["bitnet-testing-profile/runtime-profile-trace", "bitnet-testing-policy-snapshot-core/runtime-profile-trace"]
+runtime-profile-iq2s-ffi = ["bitnet-testing-profile/runtime-profile-iq2s-ffi", "bitnet-testing-policy-snapshot-core/runtime-profile-iq2s-ffi"]
+runtime-profile-cpp-ffi = ["bitnet-testing-profile/runtime-profile-cpp-ffi", "bitnet-testing-policy-snapshot-core/runtime-profile-cpp-ffi"]
+runtime-profile-fixtures = ["bitnet-testing-profile/runtime-profile-fixtures", "bitnet-testing-policy-snapshot-core/runtime-profile-fixtures"]
+runtime-profile-reporting = ["bitnet-testing-profile/runtime-profile-reporting", "bitnet-testing-policy-snapshot-core/runtime-profile-reporting"]
+runtime-profile-trend = ["bitnet-testing-profile/runtime-profile-trend", "bitnet-testing-policy-snapshot-core/runtime-profile-trend"]
+runtime-profile-integration-tests = ["bitnet-testing-profile/runtime-profile-integration-tests", "bitnet-testing-policy-snapshot-core/runtime-profile-integration-tests"]
 
 [dev-dependencies]
 proptest.workspace = true
